requirements.txt
streamlit
tensorflow
yfinance
pandas
numpy
scikit-learn
requests
plotly
python-dotenv
